refactor: 도메인 수정
- 인증 메커니즘에 맞게 null 값 가능하게 변경
